+++ test/src/test_json.cpp
@@ -312,3 +312,16 @@ TEST(test_json, nested_assignment_cause_memory_leak)
     std::cout << "now" << std::endl;
     object1["key1"] = object2["key2"];
 }
+
+TEST(test_json, memory_leak)
+{
+    bourne::json object = {"key", 1, "nested", {"key", 2}};
+
+    EXPECT_EQ(1, object["key"].to_int());
+    EXPECT_EQ(2, object["nested"]["key"].to_int());
+
+    // Create leak
+    object = object["nested"];
+
+    EXPECT_EQ(2, object["key"].to_int());
+}