+++ NEWS.rst
@@ -6,6 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Patch: Fixed bug where assignment of a json object to a contained object would
+  cause a memory leak.
 * Patch: Fixed bug where re-assignment via the ``operator[]`` would cause a memory
   leak.